backend/requirements.txt
Django==5.2.7
django-cors-headers==4.9.0
django-ninja==1.4.5
psycopg2-binary==2.9.11
pydantic==2.12.3
pydantic_core==2.41.4
celery==5.5.3
redis==7.0.1
numpy==2.2.6
pytesseract==0.3.13
opencv-python-headless==4.12.0.88
pdf2image==1.17.0
pillow==12.0.0
anthropic==0.72.0
openai==2.6.1
python-dotenv==1.2.1
tenacity==9.1.2
json_repair==0.52.4
gunicorn==23.0.0
boto3==1.35.36
django-storages[s3]==1.14.4
django-cleanup==8.1.0